refactor trip actions into trip and request action files to clarify data locations
